Don't use threads in SearchDbJob
Asyncio loops are not threadsafe so we must use a separate loop per thread. The
callbacks passed to DaemonThread need to be able update the UI, which involves
adding a task to the current loop.

As a result, we're creating tasks from the thread's loop and adding them to the
main loop. (I don't know why this doesn't ring any of asyncio's alarms, but it
should.)

This should really fix the issue where garbled escape code sequences are
sometimes inserted into the search form's query field.