Do not raise RuntimeError when send() called on finished job
Same for error() and exception().

Because there many jobs can run concurrently and it is impossible to tell their
current status, there is an unpredictable amount of edge cases in which job A
sets of some background task or process, then job B fails (which means all jobs
are finished) and THEN the background task calls some callback that calls
job.send() or .error().

It should be ok if only the first error is reported and if output may be lost in
error cases.