QueueJobBase._read_queue(): Catch exception from handle_input()
This can be an issue when we are running in a background task and exceptions are
not handled for some reason.

Without this patch, when adding a torrent to a client raises an exception,
`upsies submit ... --add-to dummy` hangs indenfinitely and nothing happens. On
the other hand, `upsies ct ... --add-to dummy` raises the exception with and
without this patch.