BbTrackerJobs: Don't make scene_check_job conditional
1. Because it should always run.

2. If it is conditional, it doesn't get focused by the TUI.

Without a condition, scene_check_job runs immediately and when the UI looks for
the next interactive job, scene_check_job can already provide a prompt. With a
condition, it only runs after release_type_job is finished. The UI looks for the
next interactive job but scene_check_job doesn't know yet because it is still
figuring things out and the UI moves on to imdb/tvmaze_job.