SceneCheckJobWidget.is_interactive: Always return `True`
The default value of `is_interactive` depends on the focusability of
`runtime_widget`.

For SceneCheckJobWidget, this is always `False` at first because we don't have
any search results yet and so we don't know if we need to prompt the user.

When we get the search results and we must prompt the user, SceneCheckJobWidget
invalidates the TUI and we end up with two focusable widgets displayed at the
same time.