download-location: Pick most promising size match for each file
For DVD images, there can be lots and lots of size matches for each file in the
torrent. This can very quickly result in millions of combinations we have to try
to successfully hash a piece.

To prevent this, we're using difflib to find the closest matching relative file
path and don't check the rest at all.