utils.http: Create new httpx.AsyncClient() object for each request
With an application-wide client object, we can't terminate the application while
requests are not finished.

We are performing requests in background tasks that we don't care about.

If AsyncClient.aclose() is called while a background task is performing a
request, it raises RuntimeError('The connection pool was closed while n HTTP
requests/responses were still in-flight.').

This seems to fix the issue.